diplom_opencart/requirements.txt
allure-pytest==2.13.5
Faker==33.3.0
jsonschema==4.18.6
jsonschema-specifications==2024.10.1
mysql-connector-python==9.2.0
outcome==1.3.0.post0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
pydentic==0.0.1.dev3
pytest==8.1.1
pytest-xdist==3.6.1
requests==2.32.3
selenium==4.19.0
webdriver-manager==4.0.2
